Move decode() into de module
